+++ backend/Tests/TestUtilities/Builders/ModuleReferenceContractLinkEventBuilder.cs
@@ -5,7 +5,7 @@ using Application.Api.GraphQL.Accounts;
 
 namespace Tests.TestUtilities.Builders;
 
-internal class ModuleReferenceContractLinkEventBuilder
+internal sealed class ModuleReferenceContractLinkEventBuilder
 {
     private ulong _blockHeight = 1;
     private readonly string _transactionHash = "";
@@ -26,9 +26,8 @@ internal class ModuleReferenceContractLinkEventBuilder
         return new ModuleReferenceContractLinkEventBuilder();
     }
 
-    internal ModuleReferenceContractLinkEvent Build()
-    {
-        return new ModuleReferenceContractLinkEvent(
+    internal ModuleReferenceContractLinkEvent Build() =>
+        new(
             _blockHeight,
             _transactionHash,
             _transactionIndex,
@@ -40,31 +39,26 @@ internal class ModuleReferenceContractLinkEventBuilder
             _action,
             _dateTimeOffset
         );
-    }
-    
-    internal ModuleReferenceContractLinkEventBuilder WithModuleReference(
-        string moduleReference)
+
+    internal ModuleReferenceContractLinkEventBuilder WithModuleReference(string moduleReference)
     {
         _moduleReference = moduleReference;
         return this;
     }
     
-    internal ModuleReferenceContractLinkEventBuilder WithBlockHeight(
-        ulong blockHeight)
+    internal ModuleReferenceContractLinkEventBuilder WithBlockHeight(ulong blockHeight)
     {
         _blockHeight = blockHeight;
         return this;
     }
     
-    internal ModuleReferenceContractLinkEventBuilder WithTransactionIndex(
-        ulong transactionIndex)
+    internal ModuleReferenceContractLinkEventBuilder WithTransactionIndex(ulong transactionIndex)
     {
         _transactionIndex = transactionIndex;
         return this;
     }
     
-    internal ModuleReferenceContractLinkEventBuilder WithEventIndex(
-        uint eventIndex)
+    internal ModuleReferenceContractLinkEventBuilder WithEventIndex(uint eventIndex)
     {
         _eventIndex = eventIndex;
         return this;
@@ -74,6 +68,12 @@ internal class ModuleReferenceContractLinkEventBuilder
         ModuleReferenceContractLinkEvent.ModuleReferenceContractLinkAction action)
     {
         _action = action;
+        return this;
+    }
+
+    public ModuleReferenceContractLinkEventBuilder WithContractAddress(ContractAddress contractAddress)
+    {
+        _contractAddress = contractAddress;
         return this;
     }
 }